Move syncing code to its own file
