+++ store/store.go
@@ -220,7 +220,7 @@ func (Clock) After(d time.Duration) <-chan time.Time {
 }
 
 func initializeORM(config Config) (*orm.ORM, error) {
-	orm, err := orm.NewORM(config.NormalizedDatabaseURL(), config.DatabaseTimeout(), config.LinkstatsURL())
+	orm, err := orm.NewORM(config.NormalizedDatabaseURL(), config.DatabaseTimeout())
 	if err != nil {
 		return nil, err
 	}
